# api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
...
  /api/plans:
    get:
      tags:
      - rest-plan
      summary: Get the information of a plan by the name
      operationId: getPlan
      parameters:
      - name: planName
        in: query
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Plan not found
          content: {}
        "200":
          description: Plan found correctly
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Plan_PlanDetails'
    put:
      tags:
      - rest-plan
      summary: Edit the information of a plan as an administrator
      operationId: editPlan
      parameters:
      - name: planName
        in: query
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditPlanRequest_PlanDetails'
        required: true
      responses:
        "404":
          description: Plan not found
          content: {}
        "403":
          description: Only access to admin account
        "200":
          description: Edited the plan correctly
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Plan_PlanDetails'
    post:
      tags:
      - rest-plan
      summary: Create a new plan as an administrator
      operationId: createPlan
      parameters:
      - name: category
        in: query
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanCreateRequest_PlanDetails'
        required: true
      responses:
        "404":
          description: There is not a category with this name
          content: {}
        "403":
          description: Only access to admin account
        "409":
          description: There is already a plan with the name given by parameter
        "201":
          description: Created the plan correctly
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Plan_PlanDetails'
  /api/plans/notFavN:
    put:
      tags:
      - rest-plan
      summary: Remove the like of a plan given its name as the logged user
      operationId: dislikePlanN
      parameters:
      - name: planName
        in: query
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Plan not found
          content: {}
        "403":
          description: Only access to registered user
        "200":
          description: Like removed correctly
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Plan_UserRegisteredPlanDetails'
  /api/plans/notFavA:
    put:
      tags:
      - rest-plan
      summary: Remove the like of a plan given the abbreviation as the logged user
      operationId: dislikePlan
      parameters:
      - name: abbrev
        in: query
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Plan not found
          content: {}
        "403":
          description: Only access to registered user
        "200":
          description: Like removed correctly
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Plan_UserRegisteredPlanDetails'
  /api/plans/favN:
    put:
      tags:
      - rest-plan
      summary: Like a plan given its name as the logged user
      operationId: likePlanN
      parameters:
      - name: planName
        in: query
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Plan not found
          content: {}
        "403":
          description: Only access to registered user
        "200":
          description: Plan liked correctly
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Plan_UserRegisteredPlanDetails'
  /api/plans/favA:
    put:
      tags:
      - rest-plan
      summary: Like a plan given the abbreviation as the logged user
      operationId: likePlan
      parameters:
      - name: abbrev
        in: query
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Plan not found
          content: {}
        "403":
          description: Only access to registered user
        "200":
          description: Plan liked correctly
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Plan_UserRegisteredPlanDetails'
...
  /api/plans/done:
    post:
      tags:
      - rest-plan
      summary: Complete the plan by name indicated as the logged user
      operationId: completePlan
      parameters:
      - name: planName
        in: query
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Plan not found
          content: {}
        "403":
          description: Only access to registered user
        "200":
          description: Completed the plan correctly
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Plan_PlanDetails'
...
  /api/plans/explore:
    get:
      tags:
      - rest-plan
      summary: Get the plan list of the indicated explore's page
      operationId: getPlansPage
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "404":
          description: Not Found
        "200":
          description: Found the plans correctly
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Plan_PlanDetails'
        "204":
          description: Page without content
  /api/plans/completedPlans:
    get:
      tags:
      - rest-plan
      summary: Get all the completed plans
      operationId: getAllCompletedPlan
      responses:
        "404":
          description: Not Found
        "200":
          description: Plans found correctly
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Plan_CompletedPlanDetails'
        "403":
          description: Only access to admin user
    delete:
      tags:
      - rest-plan
      summary: Remove a completed plan of a given user
      operationId: removeCompletedPlanbyUser
      parameters:
      - name: planName
        in: query
        required: true
        schema:
          type: string
      - name: email
        in: query
        required: true
        schema:
          type: string
      - name: date
        in: query
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Completed plan not found
          content: {}
        "403":
          description: Only access to admin user
        "200":
          description: Completed plan removed correctly
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Plan_CompletedPlanDetails'
  /api/plans/category:
    get:
      tags:
      - rest-plan
      summary: Get all the plans of a given category by category's name
      operationId: getPlansbyCategoryName
      parameters:
      - name: categoryName
        in: query
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Category not found
          content: {}
        "200":
          description: Found the plans correctly
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Plan_PlanDetails'
  /api/plans/:
    get:
      tags:
      - rest-plan
      summary: Get all the plans
      operationId: getPlans
      responses:
        "404":
          description: Not Found
        "200":
          description: Found the plans correctly
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Plan_PlanDetails'
...
components:
  schemas:
...
    EditPlanRequest_PlanDetails:
      type: object
      properties:
        newDescription:
          type: string
        abv:
          type: string
        difficulty:
          type: integer
          format: int32
    Plan_PlanDetails:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        difficulty:
          type: integer
          format: int32
        categoryName:
          type: string
    Plan_UserRegisteredPlanDetails:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        difficulty:
          type: integer
          format: int32
        likedUser:
          type: boolean
        categoryName:
          type: string
...
    PlanCreateRequest_PlanDetails:
      type: object
      properties:
        category:
          type: string
        planName:
          type: string
        abv:
          type: string
        description:
          type: string
        difficulty:
          type: integer
          format: int32
...
    Plan_CompletedPlanDetails:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        difficulty:
          type: integer
          format: int32
        categoryName:
          type: string
...